Add InheritedFleetWorkspacePermissions to GlobalRoles
This allows admins to grant permissions to all fleet workspaces except local.
FleetWorkspacePermission.ResourceRules are the rules granted in all backing namespaces for all fleet workspaces except fleet-local.
FleetWorkspacePermission.WorkspaceVerbs defines the verbs in the ClusterRole used to grant permissions to the cluster-wide fleetworkspace resources. This ClusterRole contains all workspace names except local in the resourceNames field.

Each time a fleetworkspace is created or deleted ClusterRoles and Roles will be updated to reflect that change.

When a backing ClusterRole, Role or RoleBinding is modified by an external source the GlobalRole reconciliation loop will be triggered.