Fix user impersonation in downstream clusters (#46968)
The problem was that the extras of the user doing the request were always added to the Impersonate-Extras header. This was causing an error since the impersonated user can't impersonate the extras of the user doing the request.

From now on Impersonate-Extras are passed only if the user is trying to impersonate extras (Impersonate-Extras header is present).

---------

Co-authored-by: Jonathan Crowther <jonathan.crowther@suse.com>
Co-authored-by: Peter Matseykanets <pmatseykanets@gmail.com>